+++ src/main/kotlin/kotools/types/number/StrictlyPositiveInt.kt
@@ -90,69 +90,66 @@ public value class StrictlyPositiveInt(
     public infix operator fun plus(other: Int): Int = value + other
 
     /** Adds the [other] value to this [value]. */
-    public infix operator fun plus(other: NonZeroInt): Int =
-        value + other.value
+    public infix operator fun plus(other: NonZeroInt): Int = plus(other.value)
 
     /** Adds the [other] value to this [value]. */
     public infix operator fun plus(other: PositiveInt): StrictlyPositiveInt =
-        StrictlyPositiveInt(value + other.value)
+        StrictlyPositiveInt(plus(other.value))
 
     /** Adds the [other] value to this [value]. */
-    public infix operator fun plus(other: NegativeInt): Int =
-        value + other.value
+    public infix operator fun plus(other: NegativeInt): Int = plus(other.value)
 
     /** Adds the [other] value to this [value]. */
     public infix operator fun plus(other: StrictlyNegativeInt): Int =
-        value + other.value
+        plus(other.value)
 
     /** Subtracts the [other] value from this [value]. */
     public infix operator fun minus(other: Int): Int = value - other
 
     /** Subtracts the [other] value from this [value]. */
-    public infix operator fun minus(other: NonZeroInt): Int =
-        value - other.value
+    public infix operator fun minus(other: NonZeroInt): Int = minus(other.value)
 
     /** Subtracts the [other] value from this [value]. */
     public infix operator fun minus(other: PositiveInt): Int =
-        value - other.value
+        minus(other.value)
 
     /** Subtracts the [other] value from this [value]. */
     public infix operator fun minus(other: StrictlyPositiveInt): Int =
-        value - other.value
+        minus(other.value)
 
     /** Subtracts the [other] value from this [value]. */
     public infix operator fun minus(other: NegativeInt): StrictlyPositiveInt =
-        StrictlyPositiveInt(value - other.value)
+        StrictlyPositiveInt(minus(other.value))
 
     /** Subtracts the [other] value from this [value]. */
     public infix operator fun minus(
         other: StrictlyNegativeInt
-    ): StrictlyPositiveInt = StrictlyPositiveInt(value - other.value)
+    ): StrictlyPositiveInt = StrictlyPositiveInt(minus(other.value))
 
     /** Multiplies this [value] by the [other] value. */
     public infix operator fun times(other: Int): Int = value * other
 
     /** Multiplies this [value] by the [other] value. */
     public infix operator fun times(other: NonZeroInt): NonZeroInt =
-        NonZeroInt(value * other.value)
+        NonZeroInt(times(other.value))
 
     /** Multiplies this [value] by the [other] value. */
     public infix operator fun times(other: PositiveInt): PositiveInt =
-        PositiveInt(value * other.value)
+        PositiveInt(times(other.value))
 
     /** Multiplies this [value] by the [other] value. */
     public infix operator fun times(
         other: StrictlyPositiveInt
-    ): StrictlyPositiveInt = StrictlyPositiveInt(value * other.value)
+    ): StrictlyPositiveInt = StrictlyPositiveInt(times(other.value))
 
     /** Multiplies this [value] by the [other] value. */
     public infix operator fun times(other: NegativeInt): NegativeInt =
-        NegativeInt(value * other.value)
+        NegativeInt(times(other.value))
 
     /** Multiplies this [value] by the [other] value. */
     public infix operator fun times(
         other: StrictlyNegativeInt
-    ): StrictlyNegativeInt = StrictlyNegativeInt(value * other.value)
+    ): StrictlyNegativeInt = StrictlyNegativeInt(times(other.value))
 
     /**
      * Divides this [value] by the [other] value, truncating the result to an
@@ -166,7 +163,7 @@ public value class StrictlyPositiveInt(
      * Divides this [value] by the [other] value, truncating the result to an
      * integer that is closer to zero.
      */
-    public infix operator fun div(other: NonZeroInt): Int = value / other.value
+    public infix operator fun div(other: NonZeroInt): Int = div(other.value)
 
     /**
      * Divides this [value] by the [other] value, truncating the result to an
@@ -175,14 +172,14 @@ public value class StrictlyPositiveInt(
      */
     @Throws(ArithmeticException::class)
     public infix operator fun div(other: PositiveInt): PositiveInt =
-        PositiveInt(value / other.value)
+        PositiveInt(div(other.value))
 
     /**
      * Divides this [value] by the [other] value, truncating the result to an
      * integer that is closer to zero.
      */
     public infix operator fun div(other: StrictlyPositiveInt): PositiveInt =
-        PositiveInt(value / other.value)
+        PositiveInt(div(other.value))
 
     /**
      * Divides this [value] by the [other] value, truncating the result to an
@@ -191,14 +188,14 @@ public value class StrictlyPositiveInt(
      */
     @Throws(ArithmeticException::class)
     public infix operator fun div(other: NegativeInt): NegativeInt =
-        NegativeInt(value / other.value)
+        NegativeInt(div(other.value))
 
     /**
      * Divides this [value] by the [other] value, truncating the result to an
      * integer that is closer to zero.
      */
     public infix operator fun div(other: StrictlyNegativeInt): NegativeInt =
-        NegativeInt(value / other.value)
+        NegativeInt(div(other.value))
 
     // ---------- Comparisons ----------